update converter unittests with mxnet module
